OpenXR SDK 1.0.26 (2022-11-18)
This release contains new reflection headers, fixes and improvements to
the loader and hello_xr (especially on Android), some spec
clarifications, improvements to tooling, and a variety of new vendor and
multi-vendor extensions.

-   Registry
    -   Add new XR_EXT_active_action_set_priority vendor extension.
        (internal MR 2288, internal issue 1699)
    -   Add new XR_HTC_passthrough vendor extension. (internal MR 2349)
    -   Add new XR_HTC_foveation vendor extension. (internal MR 2377)
    -   Add a warning to
        XR_COMPOSITION_LAYER_CORRECT_CHROMATIC_ABERRATION_BIT saying
        that it is not in use and planned for deprecation (internal MR
        2378, internal issue 1751)
    -   Add new XR_META_headset_id vendor extension. (internal MR 2410)
    -   Improve Schematron rules for the registry XML and update the
        tool version used. (internal MR 2418, internal MR 2426, internal
        MR 2457, internal MR 2460, internal MR 2465)
    -   Register author ID and reserve vendor extensions for ByteDance.
        (internal MR 2482, OpenXR-Docs PR 137)
    -   Register author ID for danwillm and reserve vendor extensions.
        (OpenXR-Docs PR 138)
    -   Reserve vendor extensions for Microsoft. (internal MR 2478)
    -   XR_EXTX_overlay: Fix XML markup to correct generated valid usage
        for the event structure. (internal MR 2307)
    -   XR_EXT_performance_settings: Fix XML markup to correct generated
        valid usage, bump revision. (internal MR 2306)
    -   XR_HTCX_vive_tracker_interaction: Fix XML markup to correct
        generated valid usage for the event structure. (internal MR
        2310)
    -   XR_HTC_facial_tracking: Update vendor extension to version 2.
        (internal MR 2416)
    -   specification/scripts: Added new functionality in codegen
        scripts to support creating single extension headers. Usage:
        python3 scripts/genxr.py -registry   registry/xr.xml -standaloneExtension XR_FB_color_space standalone_header
        (internal MR 2417)
-   SDK
    -   In-line comments added to openxr_reflection.h (internal MR 2357)
    -   New openxr_reflection_structs.h and
        openxr_reflection_parent_structs.h reflection headers,
        containing additional, limited reflection expansion macro
        definitions. (internal MR 2357)
    -   loader: Add missing RegCloseKey call. (internal MR 2433)
    -   loader: Report STL usage as “none” in script-built Android AAR
        because we expose no C++ symbols. (OpenXR-SDK-Source PR 332,
        internal issue 1829, internal issue 1831)
    -   loader: Minor changes to fix a missing-prototypes warning/error.
        (OpenXR-SDK-Source PR 345)
    -   hello_xr: Correctly handle the case of 0 items returned in the
        Vulkan plugin. (internal MR 2363)
    -   hello_xr: Android exit should use ANativeActivity_finish.
        (internal MR 2409, OpenXR-SDK-Source issue 329, internal issue
        1820)
    -   hello_xr: Simplify platform plugin for POSIX platforms.
        (internal MR 2443, internal MR 2436)
    -   hello_xr: Minor tidy up of initialization code. (internal MR
        2449)
    -   hello_xr: Add android.permission.VIBRATE permission needed by
        some runtimes for the controller haptics. (internal MR 2486)
    -   hello_xr: Bump Android Gradle Plugin usage to 7.0.4 to fix
        building of hello_xr on M1 device (OpenXR-SDK-Source PR 334)
    -   cmake: Use standard CMAKE_INSTALL_INCLUDEDIR to specify included
        directories. (OpenXR-SDK-Source PR 347)
    -   Android: Remove Gradle build files from loader directory: they
        were unused because the Android Gradle Plugin could not build
        our AAR file correctly as desired. (internal MR 2453)
    -   Android: Upgrade to gradle version 7. (internal MR 2474)
    -   Enable dependabot for GitHub Actions. (OpenXR-SDK-Source PR 351,
        OpenXR-SDK-Source PR 352, OpenXR-SDK-Source PR 256)
    -   Fix CI generation of NuGet packages. (OpenXR-SDK-Source PR 350)
    -   Improve GitHub CI for OpenXR-SDK-Source. (OpenXR-SDK-Source PR
        351, OpenXR-SDK-Source PR 352, OpenXR-SDK-Source PR 256)

+++ .azure-pipelines/nuget/stage_nuget.ps1
@@ -55,15 +55,17 @@ Copy-Item (Join-Path $SDKRelease "include") (Join-Path $NugetStaging "include") 
 #
 function CopyLoader($Platform) {
     $PlatformSDKPath = Join-Path $SDKRelease "$Platform"
-    $NuGetPlatformPath = Join-Path $NugetStaging "native/$Platform/release"
+    if (Test-Path $PlatformSDKPath) {
+        $NuGetPlatformPath = Join-Path $NugetStaging "native/$Platform/release"
 
-    $NugetLibPath = Join-Path $NuGetPlatformPath "lib"
-    New-Item $NugetLibPath -ItemType "directory" -Force
-    Copy-Item (Join-Path $PlatformSDKPath "lib/openxr_loader.lib") $NugetLibPath
+        $NugetLibPath = Join-Path $NuGetPlatformPath "lib"
+        New-Item $NugetLibPath -ItemType "directory" -Force
+        Copy-Item (Join-Path $PlatformSDKPath "lib/openxr_loader.lib") $NugetLibPath
 
-    $NugetBinPath = Join-Path $NuGetPlatformPath "bin"
-    New-Item $NugetBinPath -ItemType "directory" -Force
-    Copy-Item (Join-Path $PlatformSDKPath "bin/openxr_loader.dll") $NugetBinPath
+        $NugetBinPath = Join-Path $NuGetPlatformPath "bin"
+        New-Item $NugetBinPath -ItemType "directory" -Force
+        Copy-Item (Join-Path $PlatformSDKPath "bin/openxr_loader.dll") $NugetBinPath
+    }
 }
 
 # Currently there are no non-UWP ARM/ARM64 binaries available from the SDK release.
